Update User model and UsersService
This commit updates the User model and UsersService to include the email field in the User schema and add new functions for getting a user by ID, creating a user, updating a user, and deleting a user. The User model now has a required password field and the email field is set to be unique. The UsersService also includes axios calls for these new functions.

Fixes #123